+++ bootstrap3/bootstrap.py
@@ -1,9 +1,15 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from django.conf import settings
 from importlib import import_module
 
+from django import VERSION as DJANGO_VERSION
+from django.conf import settings
+
+# Do we support set_required and set_disabled?
+# See GitHub issues 337 and 345
+# TODO: Get rid of this after support for Django 1.8 LTS ends
+DBS3_SET_REQUIRED_SET_DISABLED = DJANGO_VERSION[0] < 2 and DJANGO_VERSION[1] < 10
 
 # Default settings
 BOOTSTRAP3_DEFAULTS = {
@@ -16,8 +22,7 @@ BOOTSTRAP3_DEFAULTS = {
     'include_jquery': False,
     'horizontal_label_class': 'col-md-3',
     'horizontal_field_class': 'col-md-9',
-    'set_required': True,
-    'set_disabled': False,
+
     'set_placeholder': True,
     'required_css_class': '',
     'error_css_class': 'has-error',
@@ -33,6 +38,12 @@ BOOTSTRAP3_DEFAULTS = {
         'inline': 'bootstrap3.renderers.InlineFieldRenderer',
     },
 }
+
+if DBS3_SET_REQUIRED_SET_DISABLED:
+    BOOTSTRAP3_DEFAULTS.update({
+        'set_required': True,
+        'set_disabled': False,
+    })
 
 # Start with a copy of default settings
 BOOTSTRAP3 = BOOTSTRAP3_DEFAULTS.copy()
@@ -67,7 +78,7 @@ def javascript_url():
     Return the full url to the Bootstrap JavaScript file
     """
     return get_bootstrap_setting('javascript_url') or \
-        bootstrap_url('js/bootstrap.min.js')
+           bootstrap_url('js/bootstrap.min.js')
 
 
 def css_url():
@@ -75,7 +86,7 @@ def css_url():
     Return the full url to the Bootstrap CSS file
     """
     return get_bootstrap_setting('css_url') or \
-        bootstrap_url('css/bootstrap.min.css')
+           bootstrap_url('css/bootstrap.min.css')
 
 
 def theme_url():