+++ bootstrap3/components.py
@@ -3,18 +3,22 @@ from __future__ import unicode_literals
 
 from django.forms.widgets import flatatt
 from django.utils.safestring import mark_safe
-from bootstrap3.utils import render_tag
+from bootstrap3.utils import render_tag, add_css_class
 
 from .text import text_value
 
 
-def render_icon(icon, title=''):
+def render_icon(icon, **kwargs):
     """
     Render a Bootstrap glyphicon icon
     """
     attrs = {
-        'class': 'glyphicon glyphicon-{icon}'.format(icon=icon),
+        'class': add_css_class(
+            'glyphicon glyphicon-{icon}'.format(icon=icon),
+            kwargs.get('extra_classes', ''),
+        )
     }
+    title = kwargs.get('title')
     if title:
         attrs['title'] = title
     return render_tag('span', attrs=attrs)