+++ bootstrap3/templatetags/bootstrap3.py
@@ -31,6 +31,12 @@ MESSAGE_LEVEL_CLASSES = {
     DEFAULT_MESSAGE_LEVELS.ERROR: "alert alert-danger",
 }
 
+INTEGRITY = {
+    "css": r"sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u",
+    "theme": r"sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp",
+    "javascript": r"sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa",
+}
+
 register = template.Library()
 
 
@@ -189,8 +195,12 @@ def bootstrap_css():
 
         {% bootstrap_css %}
     """
-    urls = [url for url in [bootstrap_css_url(), bootstrap_theme_url()] if url]
-    return mark_safe(''.join([render_link_tag(url) for url in urls]))
+    rendered_urls = render_link_tag(
+        bootstrap_css_url(), integrity=INTEGRITY['css'])
+    if bootstrap_theme_url():
+        rendered_urls.append(
+            render_link_tag(bootstrap_css_url(), integrity=INTEGRITY['theme']))
+    return mark_safe(''.join([url for url in rendered_urls]))
 
 
 @register.simple_tag
@@ -234,7 +244,11 @@ def bootstrap_javascript(jquery=None):
             javascript += render_tag('script', attrs={'src': url})
     url = bootstrap_javascript_url()
     if url:
-        javascript += render_tag('script', attrs={'src': url})
+        attrs = {'src': url}
+        if INTEGRITY['javascript']:
+            attrs['integrity'] = INTEGRITY['javascript']
+            attrs['crossorigin'] = 'anonymous'
+        javascript += render_tag('script', attrs=attrs)
     return mark_safe(javascript)
 
 
@@ -310,6 +324,9 @@ def bootstrap_form(*args, **kwargs):
         form
             The form that is to be rendered
 
+        exclude
+            A list of field names (comma separated) that should not be rendered
+            E.g. exclude=subject,bcc
 
         See bootstrap_field_ for other arguments
 
@@ -408,12 +425,14 @@ def bootstrap_field(*args, **kwargs):
 
         set_required
             When set to ``True`` and the field is required then the ``required`` attribute is set on the
-            rendered field
+            rendered field. This only works up to Django 1.8. Higher Django versions handle ``required``
+            natively.
 
             :default: ``True``
 
         set_disabled
-            When set to ``True`` then the ``disabled`` attribute is set on the rendered field.
+            When set to ``True`` then the ``disabled`` attribute is set on the rendered field. This only
+            works up to Django 1.8.  Higher Django versions handle ``required`` natively.
 
             :default: ``False``
 
@@ -426,6 +445,8 @@ def bootstrap_field(*args, **kwargs):
                 * ``'medium'``
                 * ``'large'``
 
+        placeholder
+            Sets the placeholder text of a textbox
 
         horizontal_label_class
             Class used on the label when the ``layout`` is set to ``horizontal``.
@@ -448,6 +469,26 @@ def bootstrap_field(*args, **kwargs):
             ``'<span class="glyphicon glyphicon-calendar"></span>'``
             
             See the `Bootstrap docs <http://getbootstrap.com/components/#input-groups-basic>` for more examples.
+
+        addon_before_class
+            Class used on the span when ``addon_before`` is used.
+
+            One of the following values:
+                
+                * ``'input-group-addon'``
+                * ``'input-group-btn'``
+
+            :default: ``input-group-addon``
+
+        addon_after_class
+            Class used on the span when ``addon_after`` is used.
+
+            One of the following values:
+                
+                * ``'input-group-addon'``
+                * ``'input-group-btn'``
+
+            :default: ``input-group-addon``
 
         error_css_class
             CSS class used when the field has an error
@@ -537,6 +578,9 @@ def bootstrap_button(*args, **kwargs):
             Name of an icon to render in the button's visible content. See bootstrap_icon_ for acceptable values.
 
         button_class
+            The class of button to use. If none is given, btn-default will be used.
+
+        extra_classes
             Any extra CSS classes that should be added to the button.
 
         size
@@ -586,6 +630,12 @@ def bootstrap_icon(icon, **kwargs):
 
         icon
             Icon name. See the `Bootstrap docs <http://getbootstrap.com/components/#glyphicons>`_ for all icons.
+
+        extra_classes
+            Extra CSS classes to add to the icon HTML
+
+        title
+            A title for the icon (HTML title attrivute)
 
     **Usage**::